--- src/components/register/info.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const info = () => {
-  return (
-    <div>info</div>
-  )
-}
-
-export default info
